rename team module to teamMember midule
