+++ src/modules/teamMember.module/teamMember.scene.ts
@@ -2,8 +2,8 @@ import { SceneNavigator } from "../../../SceneNavigator";
 import { SessionManager } from "../../../SessionManager";
 import { UserDb } from "../../db.utils/user.utils";
 import { Sender } from "../sender";
-import { TeamController } from "./team.controller";
-import { TeamService } from "./team.service";
+import { TeamController } from "./teamMember.controller";
+import { TeamService } from "./teamMember.service";
 
 export class TeamScene {
     public static enter() {
